Fixed GameEngine so that is uses the SPILocator, and moved data from GameEngine into it's own class called GameData
