+++ gameEngine/pom.xml
@@ -25,6 +25,11 @@
             <version>1.0-SNAPSHOT</version>
             <scope>compile</scope>
         </dependency>
+        <dependency>
+            <groupId>peters.car</groupId>
+            <artifactId>player</artifactId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
     </dependencies>
