+++ player/src/main/java/player/PlayerPlugin.java
@@ -2,6 +2,7 @@ package player;
 
 import interfaces.*;
 import abstractClasses.Entity;
+import utilities.GameData;
 import utilities.Inputs;
 import utilities.SPIlocator;
 import javax.swing.*;
@@ -25,19 +26,20 @@ public class PlayerPlugin extends Entity implements IPlugin, IDrawable, IProcess
 
 
     @Override
-    public Entity create(IGameEngine gameEngine) {
+    public Entity create(GameData gamedata) {
         Entity newPlayer;
         try {
             newPlayer = new PlayerPlugin();
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
+        gamedata.addDrawables((IDrawable) newPlayer);
+        gamedata.addProcesses((IProcessing) newPlayer);
         return newPlayer;
     }
 
-
     @Override
-    public Entity delete(IGameEngine gameEngine) {
+    public Entity delete(GameData gameEngine) {
         return null;
     }
 
@@ -51,7 +53,7 @@ public class PlayerPlugin extends Entity implements IPlugin, IDrawable, IProcess
     }
 
     @Override
-    public void process(ArrayList<Inputs> inputs, IGameEngine gameEngine) {
+    public void process(ArrayList<Inputs> inputs) {
         for (IMovement iMovement : getPlugin()){
             setPosition(iMovement.defaultMove(inputs,this));
         }
